+++ app/components/relicDraws/RelicDrawBannerContents.tsx
@@ -7,11 +7,13 @@ import { enlir, soulBreakTierOrder } from '../../data/enlir';
 import { RelicDrawBannerDetails } from '../../selectors/relicDraws';
 import { isAllSame } from '../../utils/typeUtils';
 import RelicDrawBannerTable from './RelicDrawBannerTable';
+import RelicSelectionPrefsMenu from './RelicSelectionPrefsMenu';
 
 interface Props {
   banner: RelicDrawBannerDetails;
   probabilities?: RelicDrawProbabilities;
   isAnonymous?: boolean;
+  visibleExchangeShopSelections?: Set<number>;
 }
 
 function sortRelics(relicIds: number[]) {
@@ -35,7 +37,7 @@ function sortRelics(relicIds: number[]) {
  */
 export class RelicDrawBannerContents extends React.PureComponent<Props> {
   renderSelections() {
-    const { banner, isAnonymous } = this.props;
+    const { banner, isAnonymous, visibleExchangeShopSelections } = this.props;
     if (!banner.selections || !banner.selections.length) {
       return null;
     }
@@ -47,6 +49,10 @@ export class RelicDrawBannerContents extends React.PureComponent<Props> {
         groupBySeries={true}
         allowSelect={true}
         includeAvailability={true}
+        prefsMenu={() => <RelicSelectionPrefsMenu isAnonymous={isAnonymous} />}
+        filter={(id: number) =>
+          !visibleExchangeShopSelections || visibleExchangeShopSelections.has(id)
+        }
       />
     );
   }