+++ app/components/relicDraws/RelicDrawBannerTable.tsx
@@ -45,6 +45,9 @@ interface Props {
   allowSelect?: boolean;
   getSelected?: (relicId: number) => boolean;
   onSelect?: (relicId: number, want: boolean) => void;
+
+  prefsMenu?: () => React.ReactNode;
+  filter?: (id: number) => boolean;
 }
 
 interface State {
@@ -275,19 +278,28 @@ export class RelicDrawBannerTable extends React.Component<Props, State> {
       caption = 'hide';
     }
     return (
-      <span>
+      <>
         {' '}
         (
         <a href="#" onClick={this.toggleCollapsed}>
           {caption}
         </a>
         )
-      </span>
+      </>
     );
   }
 
   render() {
-    const { title, relics, probabilities, allowCollapse, allowSelect, groupBySeries } = this.props;
+    const {
+      title,
+      relics,
+      probabilities,
+      allowCollapse,
+      allowSelect,
+      groupBySeries,
+      prefsMenu: PrefsMenu,
+      filter,
+    } = this.props;
 
     let showProbability: boolean;
     let commonProbability: number | null;
@@ -314,17 +326,30 @@ export class RelicDrawBannerTable extends React.Component<Props, State> {
     } else if (groupBySeries && relicsArray.length > 1) {
       relicsArray = _.sortBy(relicsArray, getRelicGroupRealmId);
     }
+    if (filter) {
+      relicsArray = relicsArray.map(i => i.filter(filter));
+    }
 
     const collapsed = allowCollapse && this.state.collapsed;
     const grouped = relicsArray.length > 1 && _.some(relicsArray, i => i.length > 1);
     this.lastRealm = null;
+
+    // TODO: Appearance at low screen sizes (mobile) is quite bad.
+    // Part of the problem is that we don't fix column widths, so the sr-only
+    // status column takes space.  Most of the problem is that we need to use
+    // a proper responsive table design to move effects to its own row, but
+    // that's hard.
+
     return (
-      <div className="table-responsive">
+      <div className={classNames('table-responsive', styles.container)}>
         <table className={classNames('table', styles.component, { [styles.grouped]: grouped })}>
           {this.renderColumnGroup(showProbability)}
           <thead>
             <tr className="thead-dark">
-              <th colSpan={colCount}>
+              <th
+                className={classNames(styles.sectionHead, { [styles.withPrefsMenu]: !!PrefsMenu })}
+                colSpan={colCount}
+              >
                 {title}
                 {allowCollapse && this.renderShowHideLink()}
                 {commonProbability && (
@@ -335,6 +360,7 @@ export class RelicDrawBannerTable extends React.Component<Props, State> {
                     Chance of drawing: {totalProbability.toFixed(2)}% total
                   </span>
                 )}
+                {PrefsMenu && <div className="float-right">{PrefsMenu()}</div>}
               </th>
             </tr>
             {!collapsed && this.renderColumnHeaders(showProbability)}