+++ app/selectors/characters.ts
@@ -2,6 +2,7 @@ import { createSelector } from 'reselect';
 
 import * as _ from 'lodash';
 
+import { enlir } from '../data';
 import { IState } from '../reducers';
 
 const idsToSet = (...ids: Array<number[] | undefined>) =>
@@ -27,4 +28,34 @@ export const getOwnedSoulBreaks = createSelector<
   ({ characters }: IState) => characters.soulBreaks,
   ({ characters }: IState) => (characters.vault || {}).soulBreaks,
   idsToSet,
+);
+
+export const getOwnedRelics = createSelector<
+  IState,
+  Set<number> | undefined,
+  Set<number> | undefined,
+  Set<number> | undefined
+>(
+  getOwnedLegendMateria,
+  getOwnedSoulBreaks,
+  (legendMateria: Set<number> | undefined, soulBreaks: Set<number> | undefined) => {
+    if (!legendMateria || !soulBreaks) {
+      return undefined;
+    }
+
+    const result = new Set<number>();
+
+    _.forEach(enlir.relicSoulBreaks, (sb, relicId) => {
+      if (soulBreaks.has(sb.id)) {
+        result.add(+relicId);
+      }
+    });
+    _.forEach(enlir.relicLegendMateria, (lm, relicId) => {
+      if (legendMateria.has(lm.id)) {
+        result.add(+relicId);
+      }
+    });
+
+    return result;
+  },
 );